# apps/doc-engine/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
    type: object
...
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.InternalCreateDocumentWithRecipientsResponse:
    properties:
      createdAt:
        type: string
      externalId:
        type: string
      id:
        type: string
      operationType:
        type: string
      recipients:
        items:
          $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.InternalDocumentRecipientResponse'
        type: array
      signerProvider:
        type: string
      status:
        type: string
      templateId:
        type: string
      templateVersionId:
        type: string
      transactionalId:
        type: string
      workspaceId:
        type: string
    type: object
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.InternalDocumentRecipientResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.InternalErrorResponse:
    properties:
      code:
        type: string
      details:
        items:
          type: string
        type: array
      error:
        type: string
    type: object
...
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.MyRolesResponse:
    properties:
      roles:
        items:
          $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.RoleEntry'
        type: array
    type: object
...
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.PaginatedTenantsWithRoleResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TenantWithRoleResponse'
        type: array
      pagination:
        $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.PaginationMeta'
    type: object
...
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.PaginationMeta:
    properties:
      page:
        type: integer
      perPage:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.RecordAccessRequest:
    properties:
      entityId:
        type: string
      entityType:
        type: string
    required:
    - entityId
    - entityType
    type: object
...
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.RoleEntry:
    properties:
      resourceId:
        description: null for SYSTEM, UUID for TENANT/WORKSPACE
        type: string
      role:
        description: The specific role (e.g., SUPERADMIN, TENANT_OWNER, ADMIN)
        type: string
      type:
        description: SYSTEM, TENANT, or WORKSPACE
        type: string
    type: object
...
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TenantWithRoleResponse:
    properties:
      code:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      lastAccessedAt:
        type: string
      name:
        type: string
      role:
        type: string
      settings:
        additionalProperties: true
        type: object
      updatedAt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Document Assembly System API - Template management and document generation
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Doc Engine API
  version: "1.0"
paths:
...
  /api/v1/documents/{documentId}/cancel:
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Document ID
        in: path
        name: documentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      summary: Cancel document
      tags:
      - Documents
...
  /api/v1/internal/documents/create:
    post:
      consumes:
      - application/json
      description: Creates a new document using the extension system (Mapper, Init,
        Injectors)
      parameters:
      - description: API Key for authentication
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: External ID (e.g., CRM entity ID)
        in: header
        name: X-External-ID
        required: true
        type: string
      - description: Template ID to use
        in: header
        name: X-Template-ID
        required: true
        type: string
      - description: Transactional ID for traceability
        in: header
        name: X-Transactional-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.InternalCreateDocumentWithRecipientsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.InternalErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.InternalErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.InternalErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.InternalErrorResponse'
      summary: Create document via internal API
      tags:
      - Internal
  /api/v1/me/access:
    post:
      consumes:
      - application/json
      description: Records that the user accessed a tenant or workspace for quick
        access history
      parameters:
      - description: Access details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.RecordAccessRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Access recorded
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Record resource access
      tags:
      - Me
  /api/v1/me/roles:
    get:
      consumes:
      - application/json
      description: |-
        Returns the current user's roles. Always includes system role if assigned.
        Optionally includes tenant role if X-Tenant-ID header is provided.
        Optionally includes workspace role if X-Workspace-ID header is provided.
      parameters:
      - description: Tenant ID to check role for
        in: header
        name: X-Tenant-ID
        type: string
      - description: Workspace ID to check role for
        in: header
        name: X-Workspace-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.MyRolesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get my roles
      tags:
      - Me
  /api/v1/me/tenants:
    get:
      consumes:
      - application/json
      description: Lists tenants where the user is an active member. Supports pagination
        and optional search by name/code.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: perPage
        type: integer
      - description: Search query for tenant name or code
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.PaginatedTenantsWithRoleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List my tenants with pagination and optional search
      tags:
      - Me
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"